multiclass classification
The suggested function trandforms the matrix with predictions into k
matrices following a one-vs-all principle. Next, the unsertainty
estimation is done for each class separately. In binary classification
the estimation happens for y(x)=P(Y=1|X=x). Following this logic, k
variance estimations are done, where instances of a particur class is
the category of interest (positive class) and all others are negative.
The output is a dataframe with predited class probabilities and variance
estimates for each class.